Integrate websocket handling into CabbageBot
CabbageBot is an event emitter. Emitted events (aside from `error`) are
`open`, when the bot connected to chat, `close` when the bot disconnected
and is no longer listening, and `event` when an event is reported from
chat.

Actual bot handlers need to listen to `event` to get the actual events
from the chat room.